+++ R/MizerSim-class.R
@@ -241,6 +241,77 @@ MizerSim <- function(params, t_dimnames = NA, t_max = 100, t_save = 1) {
     return(sim)
 }
 
+#' Validate MizerSim object and upgrade if necessary
+#' 
+#' Checks that the given MizerSim object is valid and upgrades it if necessary.
+#' Checks whether any abundances are non-finite and if any are found, a warning
+#' is issued and the simulation is truncated at the last time step where all
+#' results are finite.
+#' 
+#' Occasionally, during the development of new features for mizer, the
+#' \linkS4class{MizerSim} class or the \linkS4class{MizerParams} class gains
+#' extra slots. MizerSim objects created in older versions of mizer are then no
+#' longer valid in the new version because of the missing slots. You need to
+#' upgrade them with this function.
+#' 
+#' This function adds the missing slots and fills them with default values. It
+#' also calls [validParams()] to upgrade the MizerParams object inside the
+#' MizerSim object. Any object from version 0.4 onwards can be upgraded.
+#' 
+#' @inheritSection validParams Backwards compatibility
+#' 
+#' @param sim The MizerSim object to validate
+#' @return A valid MizerSim object
+#' @export
+validSim <- function(sim) {
+    assert_that(is(sim, "MizerSim"))
+    if (needs_upgrading(sim)) {
+        sim <- suppressWarnings(upgradeSim(sim))
+        warning("Your MizerSim object was created with an earlier version of mizer. You can upgrade it with `sim <- upgradeSim(sim)` where you should replace `sim` by the name of the variable that holds your MizerSim object.")
+    } else {
+        # Always validate the params object, even if no upgrading was required.
+        sim@params <- validParams(sim@params)
+    }
+    
+    if (!all(is.finite(sim@n))) {
+        inf_idx <- which(!is.finite(sim@n), arr.ind = TRUE)
+        max_t_idx <- min(inf_idx[, 1]) - 1
+        max_t <- as.numeric(dimnames(sim@n)$time[max_t_idx])
+        warning("The simulation failed to work beyond time = ", max_t)
+        # we can't use drop = FALSE because we do want to drop time dimension.
+        n <- sim@n[max_t_idx, , ]
+        dim(n) <- dim(sim@n)[2:3]
+        rates <- getRates(sim@params,
+                          n = n,
+                          n_pp = sim@n_pp[max_t_idx, ],
+                          n_other = sim@n_pp[max_t_idx, ],
+                          effort = sim@effort[max_t_idx, ],
+                          t = max_t)
+        inf_rates <- sapply(rates, function(x) any(!is.finite(x)))
+        if (any(inf_rates)) {
+            warning("The following rates failed to be finite: ",
+                    paste(names(rates)[inf_rates], collapse = ", "), ".")
+        }
+        sim <- truncateSim(sim, end_time = max_t)
+    }
+    validObject(sim)
+    sim
+}
+
+truncateSim <- function(sim, end_time) {
+    assert_that(is(sim, "MizerSim"))
+    times <- dimnames(sim@n)$time
+    if (!(end_time %in% times)) {
+        stop("end_time = ", end_time, " is not a valid time contained in the simulation.")
+    }
+    t_idx <- match(end_time, times)
+    sim@n <- sim@n[1:t_idx, , , drop = FALSE]
+    sim@n_pp <- sim@n_pp[1:t_idx, , drop = FALSE]
+    sim@n_other <- sim@n_other[1:t_idx, , drop = FALSE]
+    sim@effort <- sim@effort[1:t_idx, , drop = FALSE]
+    sim
+}
+
 #' Time series of size spectra
 #' 
 #' Fetch the simulation results for the size spectra over time.
@@ -272,7 +343,6 @@ NResource <- function(sim) {
 #' @param sim A MizerSim object
 #' @return For `finalN()`: An array (species x size) holding the consumer
 #'   number densities at the end of the simulation
-#' @seealso [idxFinalT()]
 #' @export
 #' @examples
 #' str(finalN(NS_sim))
@@ -297,10 +367,8 @@ finalNResource <- function(sim) {
     sim@n_pp[dim(sim@n_pp)[[1]], ]
 }
 
-#' Time index at end of simulation
-#' 
-#' @param sim A MizerSim object
-#' @return An integer giving the index for extracting the
+#' @rdname finalN
+#' @return For `idxFinalT()`: An integer giving the index for extracting the
 #'   results for the final time step
 #' @export
 #' @examples